加载报表默认宽度

+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/quartz/job/AsyncAListDateJob.java
@@ -6,6 +6,7 @@ import org.jeecg.common.system.vo.DictModel;
 import org.jeecg.common.util.dynamic.db.DynamicDBUtil;
 import org.jeecg.config.LogUtil;
 import org.jeecg.config.StringUtil;
+import org.jeecg.config.ThreadUtil;
 import org.jeecg.modules.system.service.ISysDictService;
 import org.quartz.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -55,7 +56,8 @@ public class AsyncAListDateJob implements Job {
     private void updateShares() {
         List<Object[]> sharesParams = new ArrayList<>(200);
         List<Object[]> deleteStorages = new ArrayList<>(200);
-        String query = "select id,name,links,resource_type,driver,mount_path from alist_shares where resource_type='' or resource_type is null or driver='' or driver is null or instr(mount_path,resource_type)<=0" + debug;
+        List<Map<String, Object>> removeList = new ArrayList<>();
+        String query = "select id,name,links,resource_type,driver,mount_path from alist_shares where resource_type='' or resource_type is null or driver='' or driver is null or instr(mount_path,resource_type)<=0 limit 5000";
         List<Map<String, Object>> result = DynamicDBUtil.findList(dbKey, query);
         for (Map map : result) {
             boolean isUpdate = false;
@@ -79,6 +81,7 @@ public class AsyncAListDateJob implements Job {
                     sharesParams.add(new Object[]{name, driver, resourceType, newPath, id});
                     if (StringUtils.isNotBlank(mountPath)) {
                         deleteStorages.add(new Object[]{mountPath});
+                        removeList.add(map);
                     }
                     log.info("{}: {} -> {}", driver, name, resourceType);
                 }
@@ -88,14 +91,25 @@ public class AsyncAListDateJob implements Job {
             }
             if (sharesParams.size() > batchUpdateSize) {
                 DynamicDBUtil.batchUpdate(dbKey, "update alist_shares set name=?,driver=?,resource_type=?,mount_path=? where id=?", sharesParams);
-                DynamicDBUtil.batchUpdate(dbKey, "delete from alist_storages where mount_path=?", deleteStorages);
+                if (!removeList.isEmpty()) {
+                    ThreadUtil.execute(() -> {
+                        AListJobUtil.deleteStorage(removeList);
+                        DynamicDBUtil.batchUpdate(dbKey, "delete from alist_storages where mount_path=?", deleteStorages);
+                    });
+                }
                 log.info(String.format("updateShares: delete %d,total %d", deleteStorages.size(), result.size()));
                 sharesParams.clear();
                 deleteStorages.clear();
+                removeList.clear();
             }
         }
         DynamicDBUtil.batchUpdate(dbKey, "update alist_shares set name=?,driver=?,resource_type=?,mount_path=? where id=?", sharesParams);
-        DynamicDBUtil.batchUpdate(dbKey, "delete from alist_storages where mount_path=?", deleteStorages);
+        if (!removeList.isEmpty()) {
+            ThreadUtil.execute(() -> {
+                AListJobUtil.deleteStorage(removeList);
+                DynamicDBUtil.batchUpdate(dbKey, "delete from alist_storages where mount_path=?", deleteStorages);
+            });
+        }
         log.info(String.format("updateShares: update %d,total %d", sharesParams.size(), result.size()));
         LogUtil.endTime("AsyncAListDateJob", "updateShares");
     }
